fix(models): separate catalog normalization from credentials
